restructure module: 4 classes; add modules; move prototype and rename
