+++ fastNLP/modules/aggregation/avg_pool.py
