+++ reproduction/LSTM+self_attention_sentiment_analysis/example.py
