+++ reproduction/LSTM+self_attention_sentiment_analysis/predict.py
@@ -1,4 +1,3 @@
-import torch
 import torch.nn as nn
 import torch.nn.functional as F
