+++ reproduction/LSTM+self_attention_sentiment_analysis/prepare.py
@@ -1,5 +1,7 @@
 import pickle
+
 import Word2Idx
+
 
 def get_sets(m, n):
     """
